# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: john
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: "123456"
        type: string
      role:
        example: user
        type: string
      username:
        example: john
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
  controllers.TokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  dto.DepartmentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
...
  models.Attendance:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      date:
        type: string
      employee_id:
        type: integer
      id:
        type: integer
    type: object
  models.Department:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      employee_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      password:
        type: string
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@hrms.example.com
    name: HRMS Support
    url: http://hrms.example.com/support
  description: Human Resource Management System API built with Go (Gin framework),
    GORM, and MySQL. Supports Employee management, Attendance, Payroll, Leave Requests,
    Departments, Job Titles, and User Authentication/Authorization.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://hrms.example.com/terms/
  title: HRMS API
  version: "1.0"
paths:
  /attendance/checkin/{employee_id}:
    post:
      consumes:
      - application/json
      description: Record check-in time for an employee
      parameters:
      - description: Employee ID
        in: path
        name: employee_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Attendance'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Employee Check-In
      tags:
      - Attendance
  /attendance/checkout/{employee_id}:
    post:
      consumes:
      - application/json
      description: Record check-out time for an employee
      parameters:
      - description: Employee ID
        in: path
        name: employee_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Attendance'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Employee Check-Out
      tags:
      - Attendance
  /departments:
    get:
      consumes:
      - application/json
      description: Get paginated list of departments
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DepartmentListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List Departments
      tags:
      - Department
    post:
      consumes:
      - application/json
      description: Create a new department
      parameters:
      - description: Department data
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Department
      tags:
      - Department
  /departments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete department by ID
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete Department
      tags:
      - Department
    get:
      consumes:
      - application/json
      description: Get department by ID
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Department'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Department
      tags:
      - Department
    put:
      consumes:
      - application/json
      description: Update department by ID
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated department data
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/models.Department'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update Department
      tags:
      - Department
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access & refresh tokens
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
...
  /refresh:
    post:
      description: Generate new access and refresh tokens using refresh token (Token
        Rotation)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Refresh access token
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - Auth
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by your JWT token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"